src/main/java/com/example/order/cafe/domain/OperationTimePerDay.java
package com.example.order.cafe.domain;

import com.example.order.cafe.errorMsg.OperationTimePerDayErrorMsg;

public class OperationTimePerDay {

    private Days days;
    private OperationTime operationTime;

    public OperationTimePerDay(Days days, OperationTime operationTime){
        validation(operationTime);
        this.days = days;
        this.operationTime = operationTime;
    }

    public void validation(OperationTime operationTime){
        if(operationTime == null){
            throw new IllegalArgumentException(OperationTimePerDayErrorMsg.OPERATION_TIME_NON_NULL_ERROR_MSG.getValue());
        }
    }

    public Days getDays(){
        return this.days;
    }

    public OperationTime getOperationTime(){
        return this.operationTime;
    }
}
